refactor: move NullableCoder to separate mod
