+++ sdks/rust/src/coders/standard_coders.rs
@@ -27,6 +27,9 @@
 //! The formal specifications for these coders can be found in
 //! model/pipeline/src/main/proto/beam_runner_api.proto
 
+mod nullable_coder;
+pub use nullable_coder::NullableCoder;
+
 use std::fmt;
 use std::io::{self, ErrorKind, Read, Write};
 
@@ -151,64 +154,5 @@ impl<N: fmt::Debug + VarInt + ElemType> fmt::Debug for VarIntCoder<N> {
         o.debug_struct("VarIntCoder")
             .field("urn", &Self::URN)
             .finish()
-    }
-}
-
-pub struct NullableCoder<E: ElemType> {
-    _e: std::marker::PhantomData<E>,
-}
-
-impl<E> CoderUrn for NullableCoder<E>
-where
-    E: ElemType,
-{
-    const URN: &'static str = NULLABLE_CODER_URN;
-}
-
-impl<E> Coder for NullableCoder<E>
-where
-    E: ElemType,
-{
-    fn encode(
-        &self,
-        _element: &dyn ElemType,
-        _writer: &mut dyn Write,
-        _context: &Context,
-    ) -> Result<usize, io::Error> {
-        todo!()
-    }
-
-    fn decode(
-        &self,
-        _reader: &mut dyn Read,
-        _context: &Context,
-    ) -> Result<Box<dyn ElemType>, io::Error> {
-        todo!()
-    }
-
-    fn component_coder_urns() -> Vec<super::CoderUrnTree> {
-        vec![]
-    }
-}
-
-impl<E> fmt::Debug for NullableCoder<E>
-where
-    E: ElemType,
-{
-    fn fmt(&self, o: &mut fmt::Formatter<'_>) -> std::fmt::Result {
-        o.debug_struct("NullableCoder")
-            .field("urn", &Self::URN)
-            .finish()
-    }
-}
-
-impl<E> Default for NullableCoder<E>
-where
-    E: ElemType,
-{
-    fn default() -> Self {
-        Self {
-            _e: std::marker::PhantomData,
-        }
     }
 }
